requirements.txt
langchain
langchain_community
python-dotenv
streamlit
langchain_core
langchain-groq